+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Added Czech translation
 - Added drag-and-drop to move messages to another folder
 - Added cross account copy/move mail - [#108](https://github.com/hkdb/aerion/issues/108)
+- Added draggable recipients in composer - [#111](https://github.com/hkdb/aerion/issues/111)
+- Added auto-commit recipient on lost focus - [#85](https://github.com/hkdb/aerion/issues/85)
 - Added composer del/backspace guard to prevent accidental message delete
 - Fixed detached composer system theme detection - [#153](https://github.com/hkdb/aerion/issues/153)
 - Fixed launch flow - [#154](https://github.com/hkdb/aerion/issues/154)